Make GodMode9 booting instruction clearer
A few people have been confused about how to load GodMode9 in Section VI, and have accidentally pressed the button too late or tried to time the START button, etc. The fool-proof explanation on how to load GodMode9 or the chainloader menu appears to be to press and hold START, and keep holding the START button when powering on the console.

Section VI, step 2 makes this clearer.